Corrected types in RWImprovementOccupancyValue. Updated to v2.1.15.

+++ RealWare.Core/RealWare.Core/ExternalApproach/Models/RWImprovementOccupancyValue.cs
@@ -4,8 +4,8 @@
     {
         public string AccountNo { get; set; }
         public int ImpNo { get; set; }
-        public int OccCode { get; set; }
-        public int ImpAbstractValue { get; set; }
-        public int DetailID { get; set; }
+        public long OccCode { get; set; }
+        public double ImpAbstractValue { get; set; }
+        public long DetailID { get; set; }
     }
 }
